Separated object and association serialization.
Moved object serialization into Toy::Object and included association
serialization in Toy::Store.
