+++ spec/toy/serialization_spec.rb
@@ -1,7 +1,7 @@
 require 'helper'
 
 describe Toy::Serialization do
-  uses_constants('User', 'Game', 'Move')
+  uses_objects('User', 'Move')
 
   before do
     User.attribute :name, String
@@ -28,7 +28,6 @@ describe Toy::Serialization do
         'age' => 28
       }
     }
-
   end
 
   it "correctly serializes methods" do
@@ -63,99 +62,6 @@ describe Toy::Serialization do
     user = User.new
     json = user.to_json(:except => 'id')
     MultiJson.load(json)['user'].should_not have_key('id')
-  end
-
-  describe "serializing relationships" do
-    before do
-      User.list :games, :inverse_of => :user
-      Game.reference :user
-    end
-
-    it "should include references" do
-      user = User.create(:name => 'John', :age => 28)
-      game = user.games.create
-
-      MultiJson.load(game.to_json(:include => [:user])).should == {
-        'game' => {
-          'id'      => game.id,
-          'user_id' => user.id,
-          'user'    => {
-            'name'     => 'John',
-            'game_ids' => [game.id],
-            'id'       => user.id,
-            'age'      => 28,
-          }
-        }
-      }
-    end
-
-    it "should include lists" do
-      user = User.create(:name => 'John', :age => 28)
-      game = user.games.create
-      MultiJson.load(user.to_json(:include => [:games])).should == {
-        'user' => {
-          'name'     => 'John',
-          'game_ids' => [game.id],
-          'id'       => user.id,
-          'age'      => 28,
-          'games'    => [{'id' => game.id, 'user_id' => user.id}],
-        }
-      }
-    end
-
-    it "should not cause circular reference JSON errors for references" do
-      user = User.create(:name => 'John', :age => 28)
-      game = user.games.create
-
-      MultiJson.load(ActiveSupport::JSON.encode(game.user)).should == {
-        'user' => {
-          'name'     => 'John',
-          'game_ids' => [game.id],
-          'id'       => user.id,
-          'age'      => 28
-        }
-      }
-    end
-
-    it "should not cause circular reference JSON errors for references when called indirectly" do
-      user = User.create(:name => 'John', :age => 28)
-      game = user.games.create
-
-      MultiJson.load(ActiveSupport::JSON.encode([game.user])).should == [
-        'user' => {
-          'name'     => 'John',
-          'game_ids' => [game.id],
-          'id'       => user.id,
-          'age'      => 28
-        }
-      ]
-    end
-
-    it "should not cause circular reference JSON errors for lists" do
-      user = User.create(:name => 'John', :age => 28)
-      game = user.games.create
-
-      MultiJson.load(ActiveSupport::JSON.encode(user.games)).should ==  [{
-        'game' => {
-          'id'      => game.id,
-          'user_id' => user.id
-        }
-      }]
-    end
-
-    it "should not cause circular reference JSON errors for lists when called indirectly" do
-      user = User.create(:name => 'John', :age => 28)
-      game = user.games.create
-
-      MultiJson.load(ActiveSupport::JSON.encode({:games => user.games})).should ==  {
-        'games' => [{
-          'game' => {
-            'id'      => game.id,
-            'user_id' => user.id
-          }
-        }]
-      }
-    end
   end
 
   describe "serializing specific attributes" do
@@ -228,16 +134,15 @@ describe Toy::Serialization do
   describe "#serializable_hash" do
     context "with method that is another toystore object" do
       before do
-        Game.reference(:creator, User)
-        @game = Game.create(:creator => User.create)
+        Move.class_eval { attr_accessor :creator }
       end
-      let(:game) { @game }
+
+      let(:move) { Move.new(:creator => User.new) }
 
       it "returns serializable hash of object" do
-        game.serializable_hash(:methods => [:creator]).should == {
-          'id'         => game.id,
-          'creator_id' => game.creator_id,
-          'creator'    => {'id' => game.creator.id}
+        move.serializable_hash(:methods => [:creator]).should == {
+          'id'         => move.id,
+          'creator'    => {'id' => move.creator.id}
         }
       end
     end
